@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,12 +28,12 @@ dependencies:
   intl: ^0.16.1
   charts_flutter: ^0.9.0
   webview_flutter: ^0.3.20
-  auto_size_text: ^2.1.0
   firebase_auth: ^0.16.0
   google_sign_in: ^4.4.4
   provider: ^4.0.5
   equatable: ^1.1.1
   flutter_bloc: ^3.2.0
+  path_provider: ^1.6.7
   # rxdart: ^0.24.0
 
 
