@@ -1,5 +1,6 @@
 
 
+import 'package:biometric_auth/backend/data/dataModel.dart';
 import 'package:equatable/equatable.dart';
 
 abstract class AuthenticationEvent extends Equatable {
@@ -12,14 +13,19 @@ abstract class AuthenticationEvent extends Equatable {
 //App started
 class AppStarted extends AuthenticationEvent {}
 
-//Other methods
-class OtherMethods extends AuthenticationEvent{}
 
 //LoggedIn
-class LoggedIn extends AuthenticationEvent {}
+class LoggedIn extends AuthenticationEvent {
+  final Login user;
+
+  LoggedIn(this.user);
+
+}
 
 //LoggedIn sin email
 class LoggedInWithOutEmail extends AuthenticationEvent {}
 
 //LoggedOut
-class LoggedOut extends AuthenticationEvent {}
+class LoggedOut extends AuthenticationEvent {
+  
+}
