@@ -1,86 +1,25 @@
-import 'package:meta/meta.dart';
+import 'package:biometric_auth/backend/data/dataModel.dart';
+import 'package:equatable/equatable.dart';
 
-class LoginState {
-  final bool isEmailValid;
-  final bool isPasswordValid;
-  final bool isSubmitting;
-  final bool isSuccess;
-  final bool isFailure;
-  bool get isFormValid => isEmailValid && isPasswordValid;
 
-  LoginState(
-      {@required this.isEmailValid,
-      @required this.isPasswordValid,
-      @required this.isSubmitting,
-      @required this.isSuccess,
-      @required this.isFailure});
-    // - Vacio, empty
-  factory LoginState.empty() {
-    return LoginState(
-        isEmailValid: true,
-        isPasswordValid: true,
-        isSubmitting: false,
-        isSuccess: false,
-        isFailure: false);
-  }
-  factory LoginState.loading() {
-    return LoginState(
-        isEmailValid: true,
-        isPasswordValid: true,
-        isSubmitting: true,
-        isSuccess: false,
-        isFailure: false);
-  }
-
-  factory LoginState.failure() {
-    return LoginState(
-        isEmailValid: true,
-        isPasswordValid: true,
-        isSubmitting: false,
-        isSuccess: false,
-        isFailure: true);
-  }
-
-  factory LoginState.success() {
-    return LoginState(
-        isEmailValid: true,
-        isPasswordValid: true,
-        isSubmitting: false,
-        isSuccess: true,
-        isFailure: false);
-  }
-  LoginState copyWith(
-      {bool isEmailValid,
-      bool isPasswordValid,
-      bool isSubmitting,
-      bool isSuccess,
-      bool isFailure}) {
-    return LoginState(
-        isEmailValid: isEmailValid ?? this.isEmailValid,
-        isPasswordValid: isPasswordValid ?? this.isPasswordValid,
-        isSubmitting: isSubmitting ?? this.isSubmitting,
-        isSuccess: isSuccess ?? this.isSuccess,
-        isFailure: isFailure ?? this.isFailure);
-  }
-
-  LoginState update({bool isEmailValid, bool isPasswordValid}) {
-    return copyWith(
-        isEmailValid: isEmailValid,
-        isPasswordValid: isPasswordValid,
-        isSubmitting: false,
-        isSuccess: false,
-        isFailure: false);
-  }
-
+class LoginState extends Equatable {
   @override
-  String toString() {
-    return ''' LoginState {
-      isEmailValid: $isEmailValid,
-      isPasswordValid: $isPasswordValid,
-      isSubmitting: $isSubmitting,
-      isSuccess: $isSuccess,
-      isFailure: $isFailure
-    }
-    ''';
-  }
+  List<Object> get props => [];
+}
+
+// Cuatro estados:
+// - Cargando, loading
+class Loading extends LoginState {}
+
+// - Falla, failure
+class Failure extends LoginState {}
+
+// - Exito, success
+class Success extends LoginState {
+  final Login user;
+
+  Success(this.user);
+}
+class LogOutSuccess extends LoginState {
+
 }
